+++ db/migrate/20180222142053_add_tos_acceptance_to_users.rb
@@ -1,0 +1,5 @@
+class AddTosAcceptanceToUsers < ActiveRecord::Migration[5.1]
+  def change
+    add_column :users, :tos_accepted, :datetime
+  end
+end