+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180216182524) do
+ActiveRecord::Schema.define(version: 20180222142053) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -169,6 +169,7 @@ ActiveRecord::Schema.define(version: 20180216182524) do
     t.integer "class_year"
     t.integer "old_draw_id"
     t.integer "room_id"
+    t.datetime "tos_accepted"
     t.index ["draw_id"], name: "index_users_on_draw_id"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true